polyfill: Rewrote connection infrastructure via iframe.
This drastically simplifies the communication protocol between the
requester and handler. Now we pass just a single message port through
from the requester FG to handler SW (via iframe), and those two are
entangled.

The iframe does not mediate the subsequent message passing (other than
the initial handshake), and in fact it can be destroyed after that.