Added proxy, as intermediary between requester and handler.
The new "proxy" site is designed to be hosted at a fixed, public web
address, which all handlers and requesters using the polyfill will talk
to. The proxy will track all registered handlers, and provide the
chooser UI.

polyfill: The requester is no longer required to specify a hard-coded
handler URL. (Rather, the proxy now contains a hard-coded handler URL,
but this is temporary.) The polyfill loads the proxy inside the iframe,
instead of the handler.